+++ Bundle/AdminBundle/tests/Functional/app/config/bundles.php
@@ -5,6 +5,7 @@ return [
     Doctrine\Bundle\DoctrineBundle\DoctrineBundle::class => ['all' => true],
     Doctrine\Bundle\FixturesBundle\DoctrineFixturesBundle::class => ['all' => true],
     Symfony\Bundle\SecurityBundle\SecurityBundle::class => ['all' => true],
+    \Stof\DoctrineExtensionsBundle\StofDoctrineExtensionsBundle::class => ['all' => true],
     \Symfony\Bundle\TwigBundle\TwigBundle::class => ['all' => true],
     \Umbrella\CoreBundle\UmbrellaCoreBundle::class => ['all' => true],
     \Umbrella\AdminBundle\UmbrellaAdminBundle::class => ['all' => true]